requirements.txt
requests
alembic
black
SQLAlchemy
streamlit
python-dotenv
psycopg2
plotly